+++ src/main/java/net/litetex/capes/Capes.java
@@ -8,20 +8,30 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.function.Consumer;
 import java.util.function.Predicate;
+import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import com.mojang.authlib.GameProfile;
+
 import net.litetex.capes.config.Config;
+import net.litetex.capes.handler.PlayerCapeHandler;
 import net.litetex.capes.handler.PlayerCapeHandlerManager;
-import net.litetex.capes.handler.TextureLoadThrottler;
+import net.litetex.capes.handler.ProfileTextureLoadThrottler;
 import net.litetex.capes.provider.CapeProvider;
 import net.litetex.capes.provider.MinecraftCapeProvider;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.network.ClientPlayNetworkHandler;
+import net.minecraft.client.util.SkinTextures;
 import net.minecraft.util.Identifier;
 
 
 public class Capes
 {
+	private static final Logger LOG = LoggerFactory.getLogger(Capes.class);
+	
 	public static final String MOD_ID = "cape-provider";
 	
 	public static final Identifier DEFAULT_ELYTRA_IDENTIFIER =
@@ -49,11 +59,14 @@ public class Capes
 	private final boolean validateProfile;
 	private final Duration loadThrottleSuppressDuration;
 	private final Map<CapeProvider, Set<Integer>> blockedProviderCapeHashes;
+	private final int playerCacheSize;
+	private final boolean useRealPlayerOnlineValidation;
 	
 	private final PlayerCapeHandlerManager playerCapeHandlerManager;
-	private final TextureLoadThrottler textureLoadThrottler;
+	private final ProfileTextureLoadThrottler profileTextureLoadThrottler;
 	private boolean shouldRefresh;
 	
+	@SuppressWarnings("checkstyle:MagicNumber")
 	public Capes(
 		final Config config,
 		final Consumer<Config> saveConfigFunc,
@@ -63,19 +76,37 @@ public class Capes
 		this.saveConfigFunc = saveConfigFunc;
 		this.allProviders = allProviders;
 		
-		this.validateProfile = !Boolean.FALSE.equals(this.config().isValidateProfile()); // Default -> true
+		// Calculate advanced/debug values
+		
+		this.validateProfile = !Boolean.FALSE.equals(this.config().isValidateProfile());
+		LOG.debug("validateProfile: {}", this.validateProfile);
+		
 		this.loadThrottleSuppressDuration = Optional.ofNullable(this.config().getLoadThrottleSuppressSec())
 			.map(Duration::ofSeconds)
 			.orElse(Duration.ofMinutes(3));
+		LOG.debug("loadThrottleSuppressDuration: {}", this.loadThrottleSuppressDuration);
+		
 		this.blockedProviderCapeHashes = Optional.ofNullable(this.config().getBlockedProviderCapeHashes())
 			.map(map -> map.entrySet()
 				.stream()
 				.filter(e -> allProviders.containsKey(e.getKey()))
 				.collect(Collectors.toMap(e -> allProviders.get(e.getKey()), Map.Entry::getValue)))
 			.orElseGet(Map::of);
+		LOG.debug("blockedProviderCapeHashes: {}x", this.blockedProviderCapeHashes.size());
+		
+		final Integer configPlayerCacheSize = this.config.getPlayerCacheSize();
+		this.playerCacheSize = configPlayerCacheSize != null
+			? Math.clamp(configPlayerCacheSize, 1, 100_000)
+			: 1000;
+		LOG.debug("playerCacheSize: {}", this.playerCacheSize);
+		
+		this.useRealPlayerOnlineValidation = Boolean.TRUE.equals(config.getUseRealPlayerOnlineValidation());
+		LOG.debug("useRealPlayerOnlineValidation: {}", this.useRealPlayerOnlineValidation);
 		
 		this.playerCapeHandlerManager = new PlayerCapeHandlerManager(this);
-		this.textureLoadThrottler = new TextureLoadThrottler(this.playerCapeHandlerManager);
+		this.profileTextureLoadThrottler = new ProfileTextureLoadThrottler(
+			this.playerCapeHandlerManager,
+			this.playerCacheSize());
 	}
 	
 	public void saveConfig()
@@ -100,14 +131,14 @@ public class Capes
 	
 	protected void refresh()
 	{
-		this.textureLoadThrottler.clearCache();
+		this.profileTextureLoadThrottler.clearCache();
 		this.playerCapeHandlerManager.clearCache();
 		
 		final ClientPlayNetworkHandler networkHandler = MinecraftClient.getInstance().getNetworkHandler();
 		if(networkHandler != null)
 		{
 			networkHandler.getPlayerList().forEach(e ->
-				this.textureLoadThrottler.loadIfRequired(e.getProfile()));
+				this.profileTextureLoadThrottler.loadIfRequired(e.getProfile()));
 		}
 	}
 	
@@ -151,13 +182,52 @@ public class Capes
 		return this.blockedProviderCapeHashes;
 	}
 	
-	public TextureLoadThrottler textureLoadThrottler()
+	public int playerCacheSize()
 	{
-		return this.textureLoadThrottler;
+		return this.playerCacheSize;
+	}
+	
+	public boolean useRealPlayerOnlineValidation()
+	{
+		return this.useRealPlayerOnlineValidation;
+	}
+	
+	public ProfileTextureLoadThrottler textureLoadThrottler()
+	{
+		return this.profileTextureLoadThrottler;
 	}
 	
 	public PlayerCapeHandlerManager playerCapeHandlerManager()
 	{
 		return this.playerCapeHandlerManager;
+	}
+	
+	public boolean overwriteSkinTextures(
+		final GameProfile profile,
+		final Supplier<SkinTextures> oldTexureSupplier,
+		final Consumer<SkinTextures> applyOverwrittenTextures)
+	{
+		final PlayerCapeHandler handler = this.playerCapeHandlerManager().getProfile(profile);
+		if(handler != null)
+		{
+			final Identifier capeTexture = handler.getCape();
+			if(capeTexture != null)
+			{
+				final SkinTextures oldTextures = oldTexureSupplier.get();
+				final Identifier elytraTexture = handler.hasElytraTexture()
+					&& this.config().isEnableElytraTexture()
+					? capeTexture
+					: Capes.DEFAULT_ELYTRA_IDENTIFIER;
+				applyOverwrittenTextures.accept(new SkinTextures(
+					oldTextures.texture(),
+					oldTextures.textureUrl(),
+					capeTexture,
+					elytraTexture,
+					oldTextures.model(),
+					oldTextures.secure()));
+				return true;
+			}
+		}
+		return false;
 	}
 }