+++ src/main/java/net/litetex/capes/handler/RealPlayerValidator.java
@@ -1,15 +1,15 @@
 package net.litetex.capes.handler;
 
+import java.util.Collections;
 import java.util.Map;
 import java.util.UUID;
-import java.util.WeakHashMap;
-import java.util.concurrent.ConcurrentHashMap;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.mojang.authlib.GameProfile;
 
+import net.litetex.capes.util.collections.MaxSizedHashMap;
 import net.minecraft.client.MinecraftClient;
 
 
@@ -17,7 +17,14 @@ public class RealPlayerValidator
 {
 	private static final Logger LOG = LoggerFactory.getLogger(RealPlayerValidator.class);
 	
-	private final Map<UUID, Boolean> cache = new ConcurrentHashMap<>(new WeakHashMap<>());
+	private final Map<UUID, Boolean> cache;
+	private final boolean useOnlineValidation;
+	
+	public RealPlayerValidator(final int playerCacheSize, final boolean useOnlineValidation)
+	{
+		this.cache = Collections.synchronizedMap(new MaxSizedHashMap<>(playerCacheSize));
+		this.useOnlineValidation = useOnlineValidation;
+	}
 	
 	public boolean isReal(final GameProfile profile)
 	{
@@ -26,23 +33,63 @@ public class RealPlayerValidator
 	
 	private boolean checkReal(final GameProfile profile)
 	{
-		final MinecraftClient client = MinecraftClient.getInstance();
-		
-		// The current player is always valid
-		final boolean real = profile.getId().equals(client.getSession().getUuidOrNull())
-			// Shortcut: Check if the name is valid
-			|| this.isValidName(profile.getName())
-			// Check if this is a real player (not a fake one create by a server)
-			// Use secure = false to utilize cache
-			&& client.getSessionService().fetchProfile(profile.getId(), false) != null;
+		final ValidityState validityState = this.determineIfInvalid(MinecraftClient.getInstance(), profile);
 		
 		LOG.debug(
-			"Determined that {}/{} is {}a real player",
+			"Determined that {}/{} is {}a real player: {}",
 			profile.getName(),
 			profile.getId(),
-			real ? "" : "NOT ");
+			validityState.isValid() ? "" : "NOT ",
+			validityState.name());
 		
-		return real;
+		return validityState.isValid();
+	}
+	
+	private ValidityState determineIfInvalid(final MinecraftClient client, final GameProfile profile)
+	{
+		// The current player is always valid
+		if(profile.getId().equals(client.getSession().getUuidOrNull()))
+		{
+			return ValidityState.SELF;
+		}
+		// Only valid players have version 4 (random generated)
+		// Some servers report players with different versions,
+		// however these are ignored as the cape provider can't match them
+		if(profile.getId().version() != 4)
+		{
+			return ValidityState.UUID_INCORRECT_VERSION;
+		}
+		if(!this.isValidName(profile.getName()))
+		{
+			return ValidityState.INVALID_NAME;
+		}
+		if(this.useOnlineValidation && !this.isValidSessionProfile(client, profile.getId()))
+		{
+			return ValidityState.ONLINE_VALIDATION_FAIL;
+		}
+		
+		return ValidityState.DEFAULT_OK;
+	}
+	
+	enum ValidityState
+	{
+		SELF(true),
+		UUID_INCORRECT_VERSION(false),
+		INVALID_NAME(false),
+		ONLINE_VALIDATION_FAIL(false),
+		DEFAULT_OK(true);
+		
+		private final boolean valid;
+		
+		ValidityState(final boolean valid)
+		{
+			this.valid = valid;
+		}
+		
+		public boolean isValid()
+		{
+			return this.valid;
+		}
 	}
 	
 	@SuppressWarnings("checkstyle:MagicNumber")
@@ -66,5 +113,20 @@ public class RealPlayerValidator
 			}
 		}
 		return true;
+	}
+	
+	private boolean isValidSessionProfile(final MinecraftClient client, final UUID id)
+	{
+		try
+		{
+			// Check if this is a real player (not a fake one create by a server)
+			// Use secure = false to utilize cache
+			return client.getSessionService().fetchProfile(id, false) != null;
+		}
+		catch(final Exception ex)
+		{
+			LOG.warn("Failed to validate player using online services", ex);
+			return true;
+		}
 	}
 }