+++ src/main/java/net/litetex/capes/provider/WynntilsProvider.java
@@ -46,7 +46,7 @@ public class WynntilsProvider implements CapeProvider
 			
 			final HttpRequest request = requestBuilder
 				// Does UserAgent blocking: https://github.com/Wynntils/athena-backend/pull/36
-				.header("User-Agent", "Wynntils Artemis\\3.0.6+MC-1.21.4 (client) FABRIC")
+				.header("User-Agent", "Wynntils Artemis\\3.1.6+MC-1.21.4 (client) FABRIC")
 				.header("Content-Type", "application/json")
 				.POST(HttpRequest.BodyPublishers.ofString(body.toString()))
 				.build();
@@ -96,5 +96,12 @@ public class WynntilsProvider implements CapeProvider
 	public String homepageUrl()
 	{
 		return "https://wynntils.com";
+	}
+	
+	@Override
+	public double rateLimitedReqPerSec()
+	{
+		// Wynntils has a very underperforming backend
+		return 4;
 	}
 }